agent/requirements.txt
fastapi
uvicorn
python-dotenv
websockets
langchain-openai
langgraph
python-multipart 